give TemperatureRanges enum its own file
