+++ dressForTheWeather/Models/TempClothingData.swift
@@ -55,41 +55,3 @@ var allClothingItems = [
     sandals,
     skirt
 ]
-
-enum TemperatureRanges: String, CaseIterable {
-    case veryCold
-    case cold
-    case sortaCold
-    case mild
-    case sortaWarm
-    case warm
-    case veryWarm
-    case hot
-    case veryHot
-    
-    var nextTemp: TemperatureRanges {
-        switch self {
-        case .veryCold: return .cold
-        case .cold: return .veryCold
-        case .sortaCold: return .cold
-        case .mild: return .sortaCold
-        case .sortaWarm: return .mild
-        case .warm: return .sortaWarm
-        case .veryWarm: return .warm
-        case .hot: return .veryWarm
-        case .veryHot: return .hot
-        }
-    }
-    
-//    func forTemp(temp: Double) -> TemperatureRanges {
-//        switch temp {
-//        case -20...15: return .veryCold
-//        case 16...35: return .cold
-//        case 36...50: return .sortaCold
-//        case 51...67: return .mild
-//        case 68...77: return .warm
-//        case 78...110: return .hot
-//        default: return .mild
-//        }
-//    }
-}
